more little things

+++ lib/serial.c
@@ -81,7 +81,7 @@ static char nibbleToHex(uint8_t c) {
 }
 
 //write an integer with radix 16
-void serial_writeint16(uint32_t data) {
+void serial_writeint16(uint64_t data) {
 	if (!data) {
 		serial_writestring("0x00");
 	} else {
